+++ docs/gallery/plot_spectra_timeseries.py
@@ -15,7 +15,7 @@ fig = plt.figure(figsize=(8, 4))
 
 dset = read_ww3("../_static/ww3file.nc")
 ds = dset.isel(site=0).spec.split(fmax=0.18)
-ds = ds.spec.oned().rename({"freq": "period"})
+ds = ds.efth.spec.oned().rename({"freq": "period"})
 ds = ds.assign_coords({"period": 1 / ds.period})
 ds.period.attrs.update({"standard_name": "sea_surface_wave_period", "units": "s"})
 p = ds.plot.contourf(